Improve workaround for Apexcharts bug
This uses a different (and arguably better) workaround for a bug in
Apexcharts which breaks the chart updating when using its `responsive`
options. Instead of force-updating the chart options when the series
is updated, this re-implements the little bit of responsive behavior
with Material UI's `useMediaQuery`.

Also avoids the bug where animations cause the chart to crash, so they were
re-enabled.